subprojects/testcrawl/requirements.txt
fastapi==0.116.1
starlette==0.47.3
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pymysql==1.1.2
pydantic-settings==2.10.1
requests==2.32.5
beautifulsoup4==4.13.5
lxml==6.0.1
trafilatura==2.0.0
readability-lxml==0.8.4.1
pytest==8.4.2
chardet==5.2.0
playwright==1.58.0
browser-use==0.12.5
apscheduler==3.10.4
